Refactor Checker
